tests/wpt/mozilla/tests/mozilla/window_resizeTo.html

<!doctype html>
<meta charset="utf-8">
<title>Verify that the resize event is fired when the window is resized (particularly in headless mode)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    async_test(function (t) {
        window.onresize = t.step_func(function () {
            assert_equals(window.innerWidth, 200);
            assert_equals(window.innerHeight, 200);
            t.done();
        });

        window.resizeTo(200, 200);
    }, "onresize");
</script>
